Moved `Transformed` to its own module
See https://github.com/PistonDevelopers/graphics/issues/884
